requirements.txt
# Core dependencies for News Location Mapper Pro
streamlit>=1.40.0
streamlit-folium>=0.25.0

# AI and Multi-Agent Systems
anthropic>=0.57.1
crewai>=0.70.0

# Mapping and Visualization
folium>=0.18.0
plotly>=5.18.0

# Data Processing
pandas>=2.2.0
geopy>=2.4.0

# Web Scraping
requests>=2.32.0
beautifulsoup4>=4.12.0

# Environment and Configuration
python-dotenv>=1.1.1

# Optional: Enhanced Features
# mcp>=1.0.0  # Uncomment if using advanced MCP features